move BossBarDrawParams into Terraria.DataStructures namespace
