+++ ExampleMod/Common/GlobalBossBars/ExampleGlobalBossBar.cs
@@ -5,6 +5,7 @@ using Terraria.ID;
 using Terraria.GameContent;
 using Terraria.ModLoader;
 using ReLogic.Graphics;
+using Terraria.DataStructures;
 
 namespace ExampleMod.Common.GlobalBossBars
 {
