+++ patches/tModLoader/Terraria/ModLoader/ModBossBar.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 using ReLogic.Content;
 using System;
 using System.Collections.Generic;
+using Terraria.DataStructures;
 using Terraria.GameContent;
 using Terraria.GameContent.UI.BigProgressBar;
